+++ engines/scumm/scumm.h
@@ -522,7 +522,7 @@ class ScummEngine : public Engine, public Common::Serializable {
 	friend class CharsetRendererClassic;
 	friend class CharsetRendererTownsClassic;
 	friend class ResourceManager;
-	friend class MacGui;
+	friend class MacGuiImpl;
 	friend class MacIndy3Gui;
 	friend class MacLoomGui;
